feat: coupon code migration file update for percentage(%) integer number from decimal

+++ database/migrations/2023_05_25_043037_create_coupons_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('coupon_code')->unique();
-            $table->decimal('discount', 8, 2);
+            $table->unsignedInteger('discount');
             $table->dateTime('expiry_date');
             $table->unsignedInteger('usage_limit')->default(0);
             $table->timestamps();
